+++ ChangeLog.md
@@ -1,3 +1,14 @@
 # Changelog for markov
 
-## Unreleased changes
+## 0.3.1
+Removed dependency on Data.Discrimination.
+
+Removed dependency on Generics.Deriving.
+
+Removed dependency on Data.Functor.Contravariant.
+
+Removed Markov.Instance.
+
+Added Markov.Generic.
+
+Tests now expect sorted output.